Change return type of Bitmap::GetSizeBytes to size_t
Use size_t instead of uint32_t for GetSizeBytes to avoid potential overflow on large images.

+++ CoreLib/Bitmap.hpp
@@ -181,7 +181,7 @@ public:
     }
 
 
-    uint32_t GetSizeBytes() const
+    size_t GetSizeBytes() const
     {
         return m_Data.size();
     }
